+++ Sources/FeatherStorage/ByteBufferSequence.swift
@@ -7,43 +7,47 @@
 
 import NIOCore
 
-public struct StorageByteBufferAsyncSequenceWrapper: Sendable, AsyncSequence {
+public struct ByteBufferSequence: Sendable, AsyncSequence {
 
     public typealias Element = ByteBuffer
 
     let buffer: ByteBuffer
+    let chunkSize: Int
 
     public init(
-        buffer: ByteBuffer
+        buffer: ByteBuffer,
+        chunkSize: Int = 32 * 1024
     ) {
         self.buffer = buffer
+        self.chunkSize = chunkSize
     }
 
     public struct AsyncIterator: AsyncIteratorProtocol {
         let buffer: ByteBuffer
+        let chunkSize: Int
         var currentIndex: Int = 0
-
+        
         public mutating func next() async -> ByteBuffer? {
             guard currentIndex < buffer.readableBytes else {
                 return nil
             }
-
             let endIndex = Swift.min(
-                currentIndex + 32 * 1024,
+                currentIndex + chunkSize,
                 buffer.readableBytes
             )
             let chunkRange = currentIndex..<endIndex
-
             var chunk = buffer
             chunk.moveReaderIndex(to: chunkRange.lowerBound)
             chunk.moveWriterIndex(to: chunkRange.upperBound)
-
             currentIndex = endIndex
             return chunk
         }
     }
 
     public func makeAsyncIterator() -> AsyncIterator {
-        AsyncIterator(buffer: buffer)
+        AsyncIterator(
+            buffer: buffer,
+            chunkSize: chunkSize
+        )
     }
 }
